requirements.txt
pandas==2.2.0
numpy>=1.26.0,<2.0.0
ccxt==4.2.19
ta-lib==0.4.28
scikit-learn==1.4.0
streamlit==1.31.0
polars==0.20.10
pyarrow==15.0.0
plotly==5.18.0
feedparser==6.0.10
textblob==0.17.1
nltk==3.8.1
requests==2.31.0
python-dotenv==1.0.1
python-binance==1.0.19
pytest==8.0.0
pytest-mock==3.12.0
pytest-cov==4.1.0
freezegun==1.4.0
responses==0.25.0
xgboost==2.0.3
prometheus_client==0.19.0
hurst==0.0.5
scipy==1.12.0
psutil==5.9.8
orjson==3.9.15
websockets==12.0
numba==0.60.0
llvmlite
msgpack==1.0.7
joblib==1.3.2
